refactor(web): hoist DRAWER_TRANSITION_MS to shared draft
Move the duplicated drawer-transition timing constant out of the acl
and forward RuleTable modules into a single _shared/draft/constants
source re-exported from the draft barrel.
