+++ web/src/pages/_shared/draft/constants.ts
@@ -1,0 +1,2 @@
+/** Duration in ms of the CSS transition on .yn-drawer — keep in sync with SCSS. */
+export const DRAWER_TRANSITION_MS = 220;
