+++ web/src/pages/modules/acl/AclPage.tsx
@@ -10,14 +10,13 @@ import type { Rule } from '../../../api/acl';
 import { ActionKind } from '../../../api/acl';
 import type { RuleItem, RuleDraft } from './types';
 import { rulesToNgItems, draftToRule } from './hooks';
-import { DRAWER_TRANSITION_MS } from './RuleTable';
 import RuleTable from './RuleTable';
 import RuleDrawer from './RuleDrawer';
 import type { RuleDrawerHandle } from './RuleDrawer';
 import YamlIO, { type ImportMode } from './YamlIO';
 import { SaveDiffModal } from './SaveDiffModal';
 import { useAclRuleCounters } from './useAclRuleCounters';
-import { AddConfigModal } from '../../_shared/draft';
+import { AddConfigModal, DRAWER_TRANSITION_MS } from '../../_shared/draft';
 import { DeleteConfigModal, BulkDeleteModal, CommandPaletteHeader } from '../../../components';
 import { usePalette } from '../../_shared/command-palette';
 import type { Command, RowAdapter } from '../../_shared/command-palette';
