+++ web/src/pages/modules/forward/ForwardPage.tsx
@@ -13,14 +13,13 @@ import { MODE_LABELS } from './types';
 import { ModeFilter } from './ModeFilter';
 import type { ModeFilterValue } from './ModeFilter';
 import { rulesToNgItems, draftToRule } from './hooks';
-import { DRAWER_TRANSITION_MS } from './RuleTable';
 import RuleTable from './RuleTable';
 import RuleDrawer from './RuleDrawer';
 import type { RuleDrawerHandle } from './RuleDrawer';
 import YamlIO from './YamlIO';
 import { SaveDiffModal } from './SaveDiffModal';
 import { useForwardRuleCounters } from './useForwardRuleCounters';
-import { AddConfigModal } from '../../_shared/draft';
+import { AddConfigModal, DRAWER_TRANSITION_MS } from '../../_shared/draft';
 import { DeleteConfigModal, BulkDeleteModal, CommandPaletteHeader } from '../../../components';
 import { usePalette } from '../../_shared/command-palette';
 import type { Command, RowAdapter } from '../../_shared/command-palette';
